% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engep.R
\name{compute_simi}
\alias{compute_simi}
\title{compute_simi}
\usage{
compute_simi(spcom, sccom, similarity)
}
\arguments{
\item{spcom}{common gene expression matrix of spatial data.}

\item{sccom}{common gene expression matrix of reference data.}

\item{similarity}{a vector indicates the ten similarity measure that are used.}
}
\value{
similarity matrix between spatial and reference data.
}
\description{
Function to caculate similarity matrix between spatial and reference data.
Distance measures (such as Manhattan distance, Canberra distance, Euclidean distance,
and phi_{s} proportionality measure) are transformed into similarity scores using
$s = 1/(1+d)$. The codes are modified from the calculateSimilarity function scclassify.
}
