% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engep.R
\name{engep_predict}
\alias{engep_predict}
\title{engep_predict}
\usage{
engep_predict(
  spa_counts,
  ref_list,
  pre_genes,
  nCpus = 6,
  simi_list = NULL,
  parallel = TRUE,
  k_list = NULL,
  get_baes = FALSE,
  get_weight = FALSE
)
}
\arguments{
\item{spa_counts}{gene expression matrix of spatial dataset (gene by cell).}

\item{ref_list}{each element in the list is a gene expression matrix (gene by cell) of an
original single reference dataset.}

\item{pre_genes}{an array contains names of genes to be predicted, if is NULL, ENGEP will
predict the intersection of unique genes of each references.}

\item{nCpus}{number of (maximum) cores to use for parallel execution, default to 6.}

\item{simi_list}{a vector indicates the ten similarity measures that are used, note that the
similarity measures used should be chosen from the ten similarity measures mentioned in the
document. Default is c("pearson",  "spearman","cosine", "jaccard","weighted_rank",
"manhattan","canberra","euclidean", "phi_s","rho_p")}

\item{parallel}{a logical value to indicate if the process should be run parallelly in multiple threads,
default to TURE.}

\item{k_list}{a list contains different values of $k$ (number of neighbors in knn),
default is (20,30,40,50).}

\item{get_baes}{a logical value to indicate whethe to return base results or not,
default is FAISE.}

\item{get_weight}{a logical value to indicate whethe to return weights of different references,
default is FAISE.}
}
\value{
the predicted expression levels of unmeasured genes.
}
\description{
Function to get expression levels of unmeasured genes predicted by ENGEP.
}
