% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engep.R
\name{ensemble_result}
\alias{ensemble_result}
\title{ensemble_result}
\usage{
ensemble_result(result_single)
}
\arguments{
\item{result_single}{a list contains base results predicted by different reference datasets, different
similarity measures and different values of k.}
}
\value{
ensembel result.
}
\description{
Function to combine base results by a weighted average ensemble method.
}
