% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engep.R
\name{imp_new}
\alias{imp_new}
\title{imp_new}
\usage{
imp_new(i, spcom, sccomlist, similist, k.list, sc_implist)
}
\arguments{
\item{i}{each time one reference expression matrix in the list is used to predict the expression.}

\item{spcom}{expression matrix of spatial data with common genes.}

\item{sccomlist}{a list contains expression matrices of different reference datasets with common genes.}

\item{similist}{a vector indicates the ten similarity measure that are used.}

\item{k.list}{a list contains different values of $k$ (number of neighbors in knn), default is (20,30,40,50).}

\item{sc_implist}{a list contains expression matrices of different reference datasets with genes
to be predicted.}
}
\value{
the predicted expression of unmeasured genes and the R-squared sore between the predicted expression
and the measured expression of training genes representing predictive power of reference.
}
\description{
Function to generate base results by using different reference datasets, different
similarity measures and different values of k.
}
