% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engep.R
\name{weight_r2}
\alias{weight_r2}
\title{weight_r2}
\usage{
weight_r2(r2_vec)
}
\arguments{
\item{r2_vec}{R-squared sore which presents the predictive power of reference data.}
}
\value{
weight converted from R-squared sore.
}
\description{
Function to convert the predictive power to weight, with a range of values between 0.1 and 0.9.
}
